+++ src/cmd/cmd_sym.mli
@@ -23,6 +23,7 @@ val handle_result :
   -> model_format:Cmd_utils.model_format
   -> model_out_file:Fpath.t option
   -> with_breadcrumbs:bool
+  -> entry_point:string option
   -> unit Symbolic.Choice.t
   -> unit Result.t