+++ test/replay/alloc.t
@@ -2,6 +2,7 @@
   owi: [ERROR] Assert failure: false
   model {
     symbol symbol_0 i32
+    entry_point main
   }
   owi: [ERROR] Reached problem!
   [13]